Legg til serialisering og deserialisering av BrevmottakerJournalpost
