+++ src/main/kotlin/no/nav/familie/klage/brev/domain/Brev.kt
@@ -1,5 +1,6 @@
 package no.nav.familie.klage.brev.domain
 
+import no.nav.familie.klage.distribusjon.domain.BrevmottakerJournalpost
 import no.nav.familie.klage.felles.domain.Fil
 import no.nav.familie.klage.felles.domain.Sporbar
 import org.springframework.data.annotation.Id
@@ -23,25 +24,3 @@ data class Brev(
 data class BrevmottakereJournalposter(
     val journalposter: List<BrevmottakerJournalpost>,
 )
-
-sealed interface BrevmottakerJournalpost {
-    val journalpostId: String
-    val distribusjonId: String?
-    fun medDistribusjonsId(distribusjonId: String): BrevmottakerJournalpost
-}
-
-data class BrevmottakerJournalpostMedIdent(
-    val ident: String,
-    override val journalpostId: String,
-    override val distribusjonId: String? = null,
-) : BrevmottakerJournalpost {
-    override fun medDistribusjonsId(distribusjonId: String) = copy(distribusjonId = distribusjonId)
-}
-
-data class BrevmottakerJournalpostUtenIdent(
-    val id: UUID,
-    override val journalpostId: String,
-    override val distribusjonId: String? = null,
-) : BrevmottakerJournalpost {
-    override fun medDistribusjonsId(distribusjonId: String) = copy(distribusjonId = distribusjonId)
-}
