+++ src/main/kotlin/no/nav/familie/klage/distribusjon/JournalførBrevTask.kt
@@ -5,15 +5,15 @@ import no.nav.familie.klage.brev.BrevService
 import no.nav.familie.klage.brev.BrevmottakerUtil.validerMinimumEnMottaker
 import no.nav.familie.klage.brev.domain.Brev
 import no.nav.familie.klage.brev.domain.Brevmottaker
-import no.nav.familie.klage.brev.domain.BrevmottakerJournalpost
-import no.nav.familie.klage.brev.domain.BrevmottakerJournalpostMedIdent
-import no.nav.familie.klage.brev.domain.BrevmottakerJournalpostUtenIdent
 import no.nav.familie.klage.brev.domain.BrevmottakerOrganisasjon
 import no.nav.familie.klage.brev.domain.BrevmottakerPersonMedIdent
 import no.nav.familie.klage.brev.domain.BrevmottakerPersonUtenIdent
 import no.nav.familie.klage.brev.domain.BrevmottakereJournalposter
 import no.nav.familie.klage.distribusjon.JournalføringUtil.mapAvsenderMottaker
 import no.nav.familie.klage.distribusjon.JournalføringUtil.mapBrevmottakerJournalpost
+import no.nav.familie.klage.distribusjon.domain.BrevmottakerJournalpost
+import no.nav.familie.klage.distribusjon.domain.BrevmottakerJournalpostMedIdent
+import no.nav.familie.klage.distribusjon.domain.BrevmottakerJournalpostUtenIdent
 import no.nav.familie.klage.felles.util.TaskMetadata.saksbehandlerMetadataKey
 import no.nav.familie.klage.personopplysninger.pdl.logger
 import no.nav.familie.kontrakter.felles.klage.BehandlingResultat
