+++ src/main/kotlin/no/nav/familie/klage/distribusjon/JournalføringUtil.kt
@@ -1,11 +1,11 @@
 package no.nav.familie.klage.distribusjon
 
 import no.nav.familie.klage.brev.domain.Brevmottaker
-import no.nav.familie.klage.brev.domain.BrevmottakerJournalpostMedIdent
-import no.nav.familie.klage.brev.domain.BrevmottakerJournalpostUtenIdent
 import no.nav.familie.klage.brev.domain.BrevmottakerOrganisasjon
 import no.nav.familie.klage.brev.domain.BrevmottakerPersonMedIdent
 import no.nav.familie.klage.brev.domain.BrevmottakerPersonUtenIdent
+import no.nav.familie.klage.distribusjon.domain.BrevmottakerJournalpostMedIdent
+import no.nav.familie.klage.distribusjon.domain.BrevmottakerJournalpostUtenIdent
 import no.nav.familie.kontrakter.felles.dokarkiv.AvsenderMottaker
 import no.nav.familie.kontrakter.felles.journalpost.AvsenderMottakerIdType
