+++ src/test/kotlin/no/nav/familie/klage/brev/BrevRepositoryTest.kt
@@ -1,12 +1,13 @@
 package no.nav.familie.klage.brev
 
 import no.nav.familie.klage.brev.domain.Brev
-import no.nav.familie.klage.brev.domain.BrevmottakerJournalpostMedIdent
 import no.nav.familie.klage.brev.domain.BrevmottakerOrganisasjon
 import no.nav.familie.klage.brev.domain.BrevmottakerPersonMedIdent
+import no.nav.familie.klage.brev.domain.BrevmottakerPersonUtenIdent
 import no.nav.familie.klage.brev.domain.Brevmottakere
 import no.nav.familie.klage.brev.domain.BrevmottakereJournalposter
 import no.nav.familie.klage.brev.domain.MottakerRolle
+import no.nav.familie.klage.distribusjon.domain.BrevmottakerJournalpostMedIdent
 import no.nav.familie.klage.felles.domain.Fil
 import no.nav.familie.klage.infrastruktur.config.OppslagSpringRunnerTest
 import no.nav.familie.klage.repository.findByIdOrThrow
@@ -16,6 +17,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
+import java.util.UUID
 
 internal class BrevRepositoryTest : OppslagSpringRunnerTest() {
 
@@ -42,7 +44,7 @@ internal class BrevRepositoryTest : OppslagSpringRunnerTest() {
     }
 
     @Test
-    internal fun `oppdaterMottakerJournalpost`() {
+    internal fun oppdaterMottakerJournalpost() {
         val brev = brev()
         brevRepository.insert(brev)
         val oppdatertJournalposter = BrevmottakereJournalposter(journalposter = listOf(brevmottakereJournalpost("id")))
@@ -61,6 +63,16 @@ internal class BrevRepositoryTest : OppslagSpringRunnerTest() {
         mottakere = Brevmottakere(
             personer = listOf(
                 BrevmottakerPersonMedIdent("ident", MottakerRolle.BRUKER, "navn"),
+                BrevmottakerPersonUtenIdent(
+                    id = UUID.randomUUID(),
+                    mottakerRolle = MottakerRolle.BRUKER,
+                    navn = "navn",
+                    adresselinje1 = "Adresselinje 1",
+                    adresselinje2 = "Adresselinje 2",
+                    postnummer = "1234",
+                    poststed = "Oslo",
+                    landkode = "NO",
+                ),
             ),
             organisasjoner = listOf(BrevmottakerOrganisasjon("orgnr", "navn", "mottaker")),
         ),
